Split Groups collection resources
