+++ app/collections/groups/helpers.js
@@ -1,0 +1,14 @@
+import Groups from './';
+import Homes from '/collections/homes';
+
+Groups.helpers({
+  'homes': function () {
+    // Get group ID
+    var groupId = this._id;
+
+    // Get all homes assigned to group
+    var homes = Homes.find({'groupId': groupId}).fetch();
+
+    return homes;
+  }
+});
