+++ app/collections/groups/schema.js
@@ -1,0 +1,7 @@
+import SimpleSchema from 'simpl-schema';
+
+export default GroupsSchema = new SimpleSchema({
+  name: {
+    type: String
+  }
+});
